Refactor app into modular package
